applications/conversational_agent/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
websockets>=12.0
python-dotenv>=1.0.1
langchain>=0.1.9
langchain-openai>=0.0.5
pydantic>=2.6.1
